backend/pyproject.toml
# Source: Doc 07 §21 — Poetry paket yönetimi
[tool.poetry]
name = "bist-robogo-backend"
version = "0.1.0"
description = "BIST İçin AI Destekli Otomatik Ticaret Platformu — Backend"
authors = ["bist-robogo team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"

# Core Web Framework
fastapi = "^0.115.0"
uvicorn = {version = "^0.34.0", extras = ["standard"]}
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
python-multipart = "^0.0.18"

# Database
sqlalchemy = {version = "^2.0.0", extras = ["asyncio"]}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
psycopg = {version = "^3.2.0", extras = ["binary"]}

# Redis
redis = {version = "^5.2.0", extras = ["hiredis"]}

# Celery
celery = {version = "^5.4.0", extras = ["redis"]}

# Auth & Security
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
pyotp = "^2.9.0"

# Data & ML
pandas = "^2.2.0"
numpy = "^2.1.0"
scikit-learn = "^1.6.0"
yfinance = "^0.2.0"

# HTTP Client
httpx = "^0.28.0"
aiohttp = "^3.11.0"
websockets = "^14.0"

# Messaging
confluent-kafka = "^2.6.0"

# Logging & Monitoring
structlog = "^24.4.0"
sentry-sdk = {version = "^2.19.0", extras = ["fastapi"]}
prometheus-fastapi-instrumentator = "^7.0.0"

# Utilities
python-dateutil = "^2.9.0"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.0"
factory-boy = "^3.3.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
pre-commit = "^4.0.0"

[tool.poetry.group.ml.dependencies]
xgboost = "^2.1.0"
lightgbm = "^4.5.0"
optuna = "^4.1.0"
mlflow = "^2.19.0"
onnxruntime = "^1.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short --cov=app --cov-report=term-missing"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
